+++ bridge/bridge.go
@@ -1,6 +1,7 @@
 package bridge
 
 import (
+	"github.com/42wim/matterbridge/bridge/api"
 	"github.com/42wim/matterbridge/bridge/config"
 	"github.com/42wim/matterbridge/bridge/discord"
 	"github.com/42wim/matterbridge/bridge/gitter"
@@ -70,6 +71,9 @@ func New(cfg *config.Config, bridge *config.Bridge, c chan config.Message) *Brid
 	case "rocketchat":
 		b.Config = cfg.Rocketchat[name]
 		b.Bridger = brocketchat.New(cfg.Rocketchat[name], bridge.Account, c)
+	case "api":
+		b.Config = cfg.Api[name]
+		b.Bridger = api.New(cfg.Api[name], bridge.Account, c)
 	}
 	return b
 }